rebase upstream origin/master at c9ae539

+++ experiments/models/gnn_ppa.py
@@ -1,16 +1,14 @@
 import torch
-from torch_geometric.nn import MessagePassing
 import torch.nn.functional as F
+from torch_geometric.nn import MessagePassing
 from torch_geometric.nn import global_mean_pool, global_add_pool, global_max_pool, GlobalAttention, Set2Set
 from torch_geometric.utils import degree
-from torch_scatter import scatter_mean
-import math
 
 
 class GINVirtual_ppa(torch.nn.Module):
 
-    def __init__(self, num_class, num_layer = 5, emb_dim = 300, dropout = 0.5, 
-                    gnn_type = 'gin', virtual_node = True, residual = False, JK = "last", graph_pooling = "mean"):
+    def __init__(self, num_class, num_layer=5, emb_dim=300, dropout=0.5,
+                 gnn_type='gin', virtual_node=True, residual=False, JK="last", graph_pooling="mean"):
         '''
             num_tasks (int): number of labels to be predicted
             virtual_node (bool): whether to add virtual node or not
@@ -30,10 +28,10 @@ class GINVirtual_ppa(torch.nn.Module):
 
         ### GNN to generate node embeddings
         if virtual_node:
-            self.gnn_node = GNN_node_Virtualnode(num_layer, emb_dim, JK = JK, dropout = dropout, residual = residual, gnn_type = gnn_type)
+            self.gnn_node = GNN_node_Virtualnode(num_layer, emb_dim, JK=JK, dropout=dropout, residual=residual,
+                                                 gnn_type=gnn_type)
         else:
-            self.gnn_node = GNN_node(num_layer, emb_dim, JK = JK, dropout = dropout, residual = residual, gnn_type = gnn_type)
-
+            self.gnn_node = GNN_node(num_layer, emb_dim, JK=JK, dropout=dropout, residual=residual, gnn_type=gnn_type)
 
         ### Pooling function to generate whole-graph embeddings
         if self.graph_pooling == "sum":
@@ -43,9 +41,11 @@ class GINVirtual_ppa(torch.nn.Module):
         elif self.graph_pooling == "max":
             self.pool = global_max_pool
         elif self.graph_pooling == "attention":
-            self.pool = GlobalAttention(gate_nn = torch.nn.Sequential(torch.nn.Linear(emb_dim, 2*emb_dim), torch.nn.BatchNorm1d(2*emb_dim), torch.nn.ReLU(), torch.nn.Linear(2*emb_dim, 1)))
+            self.pool = GlobalAttention(
+                gate_nn=torch.nn.Sequential(torch.nn.Linear(emb_dim, 2 * emb_dim), torch.nn.BatchNorm1d(2 * emb_dim),
+                                            torch.nn.ReLU(), torch.nn.Linear(2 * emb_dim, 1)))
         elif self.graph_pooling == "set2set":
-            self.pool = Set2Set(emb_dim, processing_steps = 2)
+            self.pool = Set2Set(emb_dim, processing_steps=2)
         else:
             raise ValueError("Invalid graph pooling type.")
 
@@ -58,7 +58,6 @@ class GINVirtual_ppa(torch.nn.Module):
         else:
             self.graph_pred_linear = torch.nn.Linear(self.emb_dim, self.num_class)
 
-
     def forward(self, batched_data):
         h_node = self.gnn_node(batched_data)
 
@@ -70,13 +69,6 @@ class GINVirtual_ppa(torch.nn.Module):
             return self.graph_pred_linear(h_graph)
 
 
-
-
-
-
-
-
-
 ### GIN convolution along the graph structure
 class GINConv(MessagePassing):
     def __init__(self, emb_dim):
@@ -84,16 +76,17 @@ class GINConv(MessagePassing):
             emb_dim (int): node embedding dimensionality
         '''
 
-        super(GINConv, self).__init__(aggr = "add")
+        super(GINConv, self).__init__(aggr="add")
 
-        self.mlp = torch.nn.Sequential(torch.nn.Linear(emb_dim, 2*emb_dim), torch.nn.BatchNorm1d(2*emb_dim), torch.nn.ReLU(), torch.nn.Linear(2*emb_dim, emb_dim))
+        self.mlp = torch.nn.Sequential(torch.nn.Linear(emb_dim, 2 * emb_dim), torch.nn.BatchNorm1d(2 * emb_dim),
+                                       torch.nn.ReLU(), torch.nn.Linear(2 * emb_dim, emb_dim))
         self.eps = torch.nn.Parameter(torch.Tensor([0]))
 
         self.edge_encoder = torch.nn.Linear(7, emb_dim)
 
     def forward(self, x, edge_index, edge_attr):
         edge_embedding = self.edge_encoder(edge_attr)
-        out = self.mlp((1 + self.eps) *x + self.propagate(edge_index, x=x, edge_attr=edge_embedding))
+        out = self.mlp((1 + self.eps) * x + self.propagate(edge_index, x=x, edge_attr=edge_embedding))
 
         return out
 
@@ -102,6 +95,7 @@ class GINConv(MessagePassing):
 
     def update(self, aggr_out):
         return aggr_out
+
 
 ### GCN convolution along the graph structure
 class GCNConv(MessagePassing):
@@ -118,14 +112,15 @@ class GCNConv(MessagePassing):
 
         row, col = edge_index
 
-        #edge_weight = torch.ones((edge_index.size(1), ), device=edge_index.device)
-        deg = degree(row, x.size(0), dtype = x.dtype) + 1
+        # edge_weight = torch.ones((edge_index.size(1), ), device=edge_index.device)
+        deg = degree(row, x.size(0), dtype=x.dtype) + 1
         deg_inv_sqrt = deg.pow(-0.5)
         deg_inv_sqrt[deg_inv_sqrt == float('inf')] = 0
 
         norm = deg_inv_sqrt[row] * deg_inv_sqrt[col]
 
-        return self.propagate(edge_index, x=x, edge_attr = edge_embedding, norm=norm) + F.relu(x + self.root_emb.weight) * 1./deg.view(-1,1)
+        return self.propagate(edge_index, x=x, edge_attr=edge_embedding, norm=norm) + F.relu(
+            x + self.root_emb.weight) * 1. / deg.view(-1, 1)
 
     def message(self, x_j, edge_attr, norm):
         return norm.view(-1, 1) * F.relu(x_j + edge_attr)
@@ -140,7 +135,8 @@ class GNN_node(torch.nn.Module):
     Output:
         node representations
     """
-    def __init__(self, num_layer, emb_dim, dropout = 0.5, JK = "last", residual = False, gnn_type = 'gin'):
+
+    def __init__(self, num_layer, emb_dim, dropout=0.5, JK="last", residual=False, gnn_type='gin'):
         '''
             emb_dim (int): node embedding dimensionality
             num_layer (int): number of GNN message passing layers
@@ -156,7 +152,7 @@ class GNN_node(torch.nn.Module):
         if self.num_layer < 2:
             raise ValueError("Number of GNN layers must be greater than 1.")
 
-        self.node_encoder = torch.nn.Embedding(1, emb_dim) # uniform input node embedding
+        self.node_encoder = torch.nn.Embedding(1, emb_dim)  # uniform input node embedding
 
         ###List of GNNs
         self.convs = torch.nn.ModuleList()
@@ -175,7 +171,6 @@ class GNN_node(torch.nn.Module):
     def forward(self, batched_data):
         x, edge_index, edge_attr, batch = batched_data.x, batched_data.edge_index, batched_data.edge_attr, batched_data.batch
 
-
         ### computing input node embedding
 
         h_list = [self.node_encoder(x)]
@@ -185,10 +180,10 @@ class GNN_node(torch.nn.Module):
             h = self.batch_norms[layer](h)
 
             if layer == self.num_layer - 1:
-                #remove relu for the last layer
-                h = F.dropout(h, self.dropout, training = self.training)
+                # remove relu for the last layer
+                h = F.dropout(h, self.dropout, training=self.training)
             else:
-                h = F.dropout(F.relu(h), self.dropout, training = self.training)
+                h = F.dropout(F.relu(h), self.dropout, training=self.training)
 
             if self.residual:
                 h += h_list[layer]
@@ -212,7 +207,8 @@ class GNN_node_Virtualnode(torch.nn.Module):
     Output:
         node representations
     """
-    def __init__(self, num_layer, emb_dim, dropout = 0.5, JK = "last", residual = False, gnn_type = 'gin'):
+
+    def __init__(self, num_layer, emb_dim, dropout=0.5, JK="last", residual=False, gnn_type='gin'):
         '''
             emb_dim (int): node embedding dimensionality
         '''
@@ -227,7 +223,7 @@ class GNN_node_Virtualnode(torch.nn.Module):
         if self.num_layer < 2:
             raise ValueError("Number of GNN layers must be greater than 1.")
 
-        self.node_encoder = torch.nn.Embedding(1, emb_dim) # uniform input node embedding
+        self.node_encoder = torch.nn.Embedding(1, emb_dim)  # uniform input node embedding
 
         ### set the initial virtual node embedding to 0.
         self.virtualnode_embedding = torch.nn.Embedding(1, emb_dim)
@@ -252,16 +248,19 @@ class GNN_node_Virtualnode(torch.nn.Module):
             self.batch_norms.append(torch.nn.BatchNorm1d(emb_dim))
 
         for layer in range(num_layer - 1):
-            self.mlp_virtualnode_list.append(torch.nn.Sequential(torch.nn.Linear(emb_dim, 2*emb_dim), torch.nn.BatchNorm1d(2*emb_dim), torch.nn.ReLU(), \
-                                                    torch.nn.Linear(2*emb_dim, emb_dim), torch.nn.BatchNorm1d(emb_dim), torch.nn.ReLU()))
-
+            self.mlp_virtualnode_list.append(
+                torch.nn.Sequential(torch.nn.Linear(emb_dim, 2 * emb_dim), torch.nn.BatchNorm1d(2 * emb_dim),
+                                    torch.nn.ReLU(), \
+                                    torch.nn.Linear(2 * emb_dim, emb_dim), torch.nn.BatchNorm1d(emb_dim),
+                                    torch.nn.ReLU()))
 
     def forward(self, batched_data):
 
         x, edge_index, edge_attr, batch = batched_data.x, batched_data.edge_index, batched_data.edge_attr, batched_data.batch
 
         ### virtual node embeddings for graphs
-        virtualnode_embedding = self.virtualnode_embedding(torch.zeros(batch[-1].item() + 1).to(edge_index.dtype).to(edge_index.device))
+        virtualnode_embedding = self.virtualnode_embedding(
+            torch.zeros(batch[-1].item() + 1).to(edge_index.dtype).to(edge_index.device))
 
         h_list = [self.node_encoder(x)]
         for layer in range(self.num_layer):
@@ -273,10 +272,10 @@ class GNN_node_Virtualnode(torch.nn.Module):
 
             h = self.batch_norms[layer](h)
             if layer == self.num_layer - 1:
-                #remove relu for the last layer
-                h = F.dropout(h, self.dropout, training = self.training)
+                # remove relu for the last layer
+                h = F.dropout(h, self.dropout, training=self.training)
             else:
-                h = F.dropout(F.relu(h), self.dropout, training = self.training)
+                h = F.dropout(F.relu(h), self.dropout, training=self.training)
 
             if self.residual:
                 h = h + h_list[layer]
@@ -290,9 +289,12 @@ class GNN_node_Virtualnode(torch.nn.Module):
                 ### transform virtual nodes using MLP
 
                 if self.residual:
-                    virtualnode_embedding = virtualnode_embedding + F.dropout(self.mlp_virtualnode_list[layer](virtualnode_embedding_temp), self.dropout, training = self.training)
+                    virtualnode_embedding = virtualnode_embedding + F.dropout(
+                        self.mlp_virtualnode_list[layer](virtualnode_embedding_temp), self.dropout,
+                        training=self.training)
                 else:
-                    virtualnode_embedding = F.dropout(self.mlp_virtualnode_list[layer](virtualnode_embedding_temp), self.dropout, training = self.training)
+                    virtualnode_embedding = F.dropout(self.mlp_virtualnode_list[layer](virtualnode_embedding_temp),
+                                                      self.dropout, training=self.training)
 
         ### Different implementations of Jk-concat
         if self.JK == "last":
